chore: clean up app.tsx
