+++ src/components/app.tsx
@@ -1,46 +1,22 @@
-import { Resizable } from "re-resizable";
-import { useEffect, useRef, useState } from "react";
 import shallow from "zustand/shallow";
 import useNavigationStore from "../hooks/navigation-store";
+import { useDefaultHeight } from "../hooks/use-default-height";
 import { select } from "../lib/utils/select";
 import styles from "./app.module.css";
 import EditorPanel from "./editor/editor-panel";
-import DragHandle from "./icons/drag-handle-icon";
+import Footer from "./footer";
 import { Notifications } from "./notifications/notifications";
 import Overlays from "./overlays/overlays";
 import ReloadPrompt from "./reload-prompt/reload-prompt";
 import StatusBar from "./status-bar/status-bar";
-import Terminal from "./terminal/terminal";
 import Welcome from "./welcome/welcome";
 
 function App() {
-  const [defaultHeight, setDefaultHeight] = useState<string | null>(null);
   const { panel, overlay } = useNavigationStore(
     select("panel", "overlay"),
     shallow
   );
-
-  const ref = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (!ref.current) {
-      setDefaultHeight("30vh");
-      return;
-    }
-
-    const sectionHeight = ref.current.offsetHeight + 100;
-    const spaceRemaining = window.innerHeight - sectionHeight;
-    const percentage = (spaceRemaining / window.innerHeight) * 100;
-
-    if (percentage < 30) {
-      setDefaultHeight("30vh");
-      return;
-    }
-
-    setDefaultHeight(percentage.toFixed(2) + "vh");
-  }, [ref.current]);
-
-  const handle = <DragHandle className={styles.dragHandle} />;
+  const { ref, defaultHeight } = useDefaultHeight();
 
   return (
     <div className={styles.app}>
@@ -49,21 +25,7 @@ function App() {
           {panel === "welcome" && <Welcome ref={ref} />}
           {panel === "editor" && <EditorPanel />}
         </section>
-        {defaultHeight && (
-          <Resizable
-            className={styles.footer}
-            enable={{ top: true }}
-            defaultSize={{ height: defaultHeight, width: "100%" }}
-            maxHeight="80vh"
-            minHeight="10vh"
-            handleClasses={{ top: styles.handle }}
-            handleComponent={{
-              top: handle,
-            }}
-          >
-            <Terminal />
-          </Resizable>
-        )}
+        <Footer defaultHeight={defaultHeight} />
         <StatusBar />
       </main>
       <Overlays overlay={overlay} />
